+++ spec/integration/rest/api/v2010/account/usage/trigger.spec.js
@@ -203,7 +203,7 @@ describe('Trigger', function() {
       var opts = {
         callbackUrl: 'https://example.com',
         triggerValue: 'triggerValue',
-        usageCategory: 'answering-machine-detection'
+        usageCategory: 'agent-conference'
       };
       var promise = client.api.v2010.accounts('ACXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXX')
                                     .usage
@@ -221,7 +221,7 @@ describe('Trigger', function() {
       var values = {
         CallbackUrl: 'https://example.com',
         TriggerValue: 'triggerValue',
-        UsageCategory: 'answering-machine-detection',
+        UsageCategory: 'agent-conference',
       };
       holodeck.assertHasRequest(new Request({
           method: 'POST',
@@ -256,7 +256,7 @@ describe('Trigger', function() {
       var opts = {
         callbackUrl: 'https://example.com',
         triggerValue: 'triggerValue',
-        usageCategory: 'answering-machine-detection'
+        usageCategory: 'agent-conference'
       };
       var promise = client.api.v2010.accounts('ACXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXX')
                                     .usage