+++ spec/integration/rest/taskrouter/v1/workspace/event.spec.js
@@ -70,10 +70,12 @@ describe('Event', function() {
               'worker_name': 'JustinWorker',
               'worker_sid': 'WKaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa',
               'worker_time_in_previous_activity': '26',
+              'worker_time_in_previous_activity_ms': '26123',
               'workspace_name': 'WorkspaceName',
               'workspace_sid': 'WSaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa'
           },
           'event_date': '2015-02-07T00:32:41Z',
+          'event_date_ms': 987654321111,
           'event_type': 'worker.activity',
           'resource_sid': 'WKaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa',
           'resource_type': 'worker',
@@ -81,6 +83,7 @@ describe('Event', function() {
           'sid': 'EVaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa',
           'source': 'twilio',
           'source_ip_address': '1.2.3.4',
+          'workspace_sid': 'WSaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa',
           'url': 'https://taskrouter.twilio.com/v1/Workspaces/WSaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa/Events/EVaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa'
       });
 
@@ -114,10 +117,12 @@ describe('Event', function() {
                       'worker_name': 'JustinWorker',
                       'worker_sid': 'WKaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa',
                       'worker_time_in_previous_activity': '26',
+                      'worker_time_in_previous_activity_ms': '26123',
                       'workspace_name': 'WorkspaceName',
                       'workspace_sid': 'WSaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa'
                   },
                   'event_date': '2015-02-07T00:32:41Z',
+                  'event_date_ms': 987654321111,
                   'event_type': 'worker.activity',
                   'resource_sid': 'WKaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa',
                   'resource_type': 'worker',
@@ -125,6 +130,7 @@ describe('Event', function() {
                   'sid': 'EVaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa',
                   'source': 'twilio',
                   'source_ip_address': '1.2.3.4',
+                  'workspace_sid': 'WSaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa',
                   'url': 'https://taskrouter.twilio.com/v1/Workspaces/WSaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa/Events/EVaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa'
               }
           ],
@@ -160,10 +166,12 @@ describe('Event', function() {
                       'worker_name': 'JustinWorker',
                       'worker_sid': 'WKaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa',
                       'worker_time_in_previous_activity': '26',
+                      'worker_time_in_previous_activity_ms': '26123',
                       'workspace_name': 'WorkspaceName',
                       'workspace_sid': 'WSaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa'
                   },
                   'event_date': '2015-02-07T00:32:41Z',
+                  'event_date_ms': 987654321111,
                   'event_type': 'worker.activity',
                   'resource_sid': 'WKaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa',
                   'resource_type': 'worker',
@@ -171,6 +179,7 @@ describe('Event', function() {
                   'sid': 'EVaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa',
                   'source': 'twilio',
                   'source_ip_address': '1.2.3.4',
+                  'workspace_sid': 'WSaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa',
                   'url': 'https://taskrouter.twilio.com/v1/Workspaces/WSaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa/Events/EVaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa'
               }
           ],
@@ -211,10 +220,12 @@ describe('Event', function() {
                       'worker_name': 'JustinWorker',
                       'worker_sid': 'WKaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa',
                       'worker_time_in_previous_activity': '26',
+                      'worker_time_in_previous_activity_ms': '26123',
                       'workspace_name': 'WorkspaceName',
                       'workspace_sid': 'WSaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa'
                   },
                   'event_date': '2015-02-07T00:32:41Z',
+                  'event_date_ms': 987654321111,
                   'event_type': 'worker.activity',
                   'resource_sid': 'WKaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa',
                   'resource_type': 'worker',
@@ -222,6 +233,7 @@ describe('Event', function() {
                   'sid': 'EVaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa',
                   'source': 'twilio',
                   'source_ip_address': '1.2.3.4',
+                  'workspace_sid': 'WSaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa',
                   'url': 'https://taskrouter.twilio.com/v1/Workspaces/WSaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa/Events/EVaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa'
               }
           ],
@@ -279,10 +291,12 @@ describe('Event', function() {
                       'worker_name': 'JustinWorker',
                       'worker_sid': 'WKaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa',
                       'worker_time_in_previous_activity': '26',
+                      'worker_time_in_previous_activity_ms': '26123',
                       'workspace_name': 'WorkspaceName',
                       'workspace_sid': 'WSaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa'
                   },
                   'event_date': '2015-02-07T00:32:41Z',
+                  'event_date_ms': 987654321111,
                   'event_type': 'worker.activity',
                   'resource_sid': 'WKaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa',
                   'resource_type': 'worker',
@@ -290,6 +304,7 @@ describe('Event', function() {
                   'sid': 'EVaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa',
                   'source': 'twilio',
                   'source_ip_address': '1.2.3.4',
+                  'workspace_sid': 'WSaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa',
                   'url': 'https://taskrouter.twilio.com/v1/Workspaces/WSaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa/Events/EVaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa'
               }
           ],