+++ spec/integration/rest/taskrouter/v1/workspace/taskQueue/taskQueueStatistics.spec.js
@@ -110,7 +110,7 @@ describe('TaskQueueStatistics', function() {
                   }
               ],
               'longest_task_waiting_age': 0,
-              'longest_task_waiting_sid': null,
+              'longest_task_waiting_sid': 'WTaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa',
               'tasks_by_status': {
                   'assigned': 0,
                   'pending': 0,