+++ spec/integration/rest/taskrouter/v1/workspace/taskQueue/taskQueuesStatistics.spec.js
@@ -94,7 +94,7 @@ describe('TaskQueuesStatistics', function() {
                           }
                       ],
                       'longest_task_waiting_age': 0,
-                      'longest_task_waiting_sid': null,
+                      'longest_task_waiting_sid': 'WTaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa',
                       'tasks_by_status': {
                           'assigned': 0,
                           'pending': 0,
@@ -180,7 +180,7 @@ describe('TaskQueuesStatistics', function() {
                           }
                       ],
                       'longest_task_waiting_age': 0,
-                      'longest_task_waiting_sid': null,
+                      'longest_task_waiting_sid': 'WTaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa',
                       'tasks_by_status': {
                           'assigned': 0,
                           'pending': 0,
@@ -271,7 +271,7 @@ describe('TaskQueuesStatistics', function() {
                           }
                       ],
                       'longest_task_waiting_age': 0,
-                      'longest_task_waiting_sid': null,
+                      'longest_task_waiting_sid': 'WTaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa',
                       'tasks_by_status': {
                           'assigned': 0,
                           'pending': 0,
@@ -380,7 +380,7 @@ describe('TaskQueuesStatistics', function() {
                           }
                       ],
                       'longest_task_waiting_age': 0,
-                      'longest_task_waiting_sid': null,
+                      'longest_task_waiting_sid': 'WTaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa',
                       'tasks_by_status': {
                           'assigned': 0,
                           'pending': 0,