+++ spec/integration/rest/taskrouter/v1/workspace/workflow/workflowRealTimeStatistics.spec.js
@@ -60,6 +60,7 @@ describe('WorkflowRealTimeStatistics', function() {
     function() {
       var body = JSON.stringify({
           'longest_task_waiting_age': 100,
+          'longest_task_waiting_sid': 'WTaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa',
           'url': 'https://taskrouter.twilio.com/v1/Workspaces/WSaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa/Workflows/WWaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa/RealTimeStatistics',
           'tasks_by_priority': {},
           'workspace_sid': 'WSaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa',