+++ spec/integration/rest/taskrouter/v1/workspace/workspaceStatistics.spec.js
@@ -107,7 +107,7 @@ describe('WorkspaceStatistics', function() {
                   }
               ],
               'longest_task_waiting_age': 0,
-              'longest_task_waiting_sid': null,
+              'longest_task_waiting_sid': 'WTaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa',
               'tasks_by_status': {
                   'assigned': 0,
                   'pending': 0,