+++ spec/integration/rest/verify/v1/service/verificationCheck.spec.js
@@ -65,6 +65,8 @@ describe('VerificationCheck', function() {
           'channel': 'sms',
           'status': 'approved',
           'valid': false,
+          'amount': '$29.99',
+          'payee': 'Acme',
           'date_created': '2015-07-30T20:00:00Z',
           'date_updated': '2015-07-30T20:00:00Z'
       });